LRQA-43938 Consolidate definition classes back to one class
